moved fixture to the correct file
